+++ include/roq/fbs/api.hpp
@@ -3752,7 +3752,8 @@ struct MarketByOrderUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
     VT_EXCHANGE_SEQUENCE = 18,
     VT_PRICE_DECIMALS = 20,
     VT_QUANTITY_DECIMALS = 22,
-    VT_CHECKSUM = 24
+    VT_MAX_DEPTH = 24,
+    VT_CHECKSUM = 26
   };
   uint16_t stream_id() const {
     return GetField<uint16_t>(VT_STREAM_ID, 0);
@@ -3784,6 +3785,9 @@ struct MarketByOrderUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
   roq::fbs::Decimals quantity_decimals() const {
     return static_cast<roq::fbs::Decimals>(GetField<uint8_t>(VT_QUANTITY_DECIMALS, 0));
   }
+  uint16_t max_depth() const {
+    return GetField<uint16_t>(VT_MAX_DEPTH, 0);
+  }
   uint32_t checksum() const {
     return GetField<uint32_t>(VT_CHECKSUM, 0);
   }
@@ -3805,6 +3809,7 @@ struct MarketByOrderUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
            VerifyField<int64_t>(verifier, VT_EXCHANGE_SEQUENCE, 8) &&
            VerifyField<uint8_t>(verifier, VT_PRICE_DECIMALS, 1) &&
            VerifyField<uint8_t>(verifier, VT_QUANTITY_DECIMALS, 1) &&
+           VerifyField<uint16_t>(verifier, VT_MAX_DEPTH, 2) &&
            VerifyField<uint32_t>(verifier, VT_CHECKSUM, 4) &&
            verifier.EndTable();
   }
@@ -3844,6 +3849,9 @@ struct MarketByOrderUpdateBuilder {
   void add_quantity_decimals(roq::fbs::Decimals quantity_decimals) {
     fbb_.AddElement<uint8_t>(MarketByOrderUpdate::VT_QUANTITY_DECIMALS, static_cast<uint8_t>(quantity_decimals), 0);
   }
+  void add_max_depth(uint16_t max_depth) {
+    fbb_.AddElement<uint16_t>(MarketByOrderUpdate::VT_MAX_DEPTH, max_depth, 0);
+  }
   void add_checksum(uint32_t checksum) {
     fbb_.AddElement<uint32_t>(MarketByOrderUpdate::VT_CHECKSUM, checksum, 0);
   }
@@ -3870,6 +3878,7 @@ inline flatbuffers::Offset<MarketByOrderUpdate> CreateMarketByOrderUpdate(
     int64_t exchange_sequence = 0,
     roq::fbs::Decimals price_decimals = roq::fbs::Decimals::Undefined,
     roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals::Undefined,
+    uint16_t max_depth = 0,
     uint32_t checksum = 0) {
   MarketByOrderUpdateBuilder builder_(_fbb);
   builder_.add_exchange_sequence(exchange_sequence);
@@ -3879,6 +3888,7 @@ inline flatbuffers::Offset<MarketByOrderUpdate> CreateMarketByOrderUpdate(
   builder_.add_bids(bids);
   builder_.add_symbol(symbol);
   builder_.add_exchange(exchange);
+  builder_.add_max_depth(max_depth);
   builder_.add_stream_id(stream_id);
   builder_.add_quantity_decimals(quantity_decimals);
   builder_.add_price_decimals(price_decimals);
@@ -3903,6 +3913,7 @@ inline flatbuffers::Offset<MarketByOrderUpdate> CreateMarketByOrderUpdateDirect(
     int64_t exchange_sequence = 0,
     roq::fbs::Decimals price_decimals = roq::fbs::Decimals::Undefined,
     roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals::Undefined,
+    uint16_t max_depth = 0,
     uint32_t checksum = 0) {
   auto exchange__ = exchange ? _fbb.CreateString(exchange) : 0;
   auto symbol__ = symbol ? _fbb.CreateString(symbol) : 0;
@@ -3920,6 +3931,7 @@ inline flatbuffers::Offset<MarketByOrderUpdate> CreateMarketByOrderUpdateDirect(
       exchange_sequence,
       price_decimals,
       quantity_decimals,
+      max_depth,
       checksum);
 }