+++ include/roq/fbs/encode.hpp
@@ -385,6 +385,7 @@ auto encode(B &builder, roq::MarketByOrderUpdate const &value) {
       value.exchange_sequence,
       encode(builder, value.price_decimals),
       encode(builder, value.quantity_decimals),
+      value.max_depth,
       value.checksum);
 }